pyproject.toml
[project]
name = "langchain_mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-devops>=7.1.0b4",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "langchain>=0.3.23",
    "langchain-mcp-adapters>=0.0.7",
    "langchain-openai>=0.3.12",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.3.27",
    "pydantic-settings>=2.8.1",
    "uvicorn>=0.34.0",
]
